wagtail-grapple/requirements.txt
# Test requirements.
black==19.3b0
Django>=2.2,<2.3
wagtail>=2.5,<2.7
graphene-django>=2.2.0
colorthief
channels==1.1.8
asgi_redis
graphql_ws
wagtailmedia
graphql-core==2.2.1
factory-boy==2.12.0
wagtail-factories==2.0.0
django-cors-headers==3.0.2
scour==0.37
wagtail-headless-preview==0.1.4